my_first_streamlitapp/requirements.txt
streamlit>=1.24.0
pandas>=1.5.0
plotly>=5.13.0
scikit-learn>=1.0.0
openai>=1.0.0
python-dotenv>=1.0.0